refactor - move class out
