+++ server/services/src/main/java/org/zanata/action/FrontendManifest.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2017, Red Hat, Inc. and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package org.zanata.action
+
+import org.codehaus.jackson.annotate.JsonIgnoreProperties
+import org.codehaus.jackson.annotate.JsonProperty
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+data class FrontendManifest(
+        @JsonProperty("editor.css") val editorCss: String,
+        @JsonProperty("editor.js") val editorJs: String,
+        @JsonProperty("frontend.css") val frontendCss: String,
+        @JsonProperty("frontend.js") val frontendJs: String,
+        @JsonProperty("frontend.legacy.js") val legacyJs: String,
+        // this js module is referenced in zanata-frontend/src/frontend/app/editor/index.js as npm module 'intl-polyfill'.
+        // It appears to work properly despite the hashed name.
+        @JsonProperty("intl-polyfill.js") val intlPolyFillJs: String,
+        @JsonProperty("runtime.js") val runtime: String
+)
